Added run order (if applicable, the unit will use the run animation and speed for movement, otherwise walk is used).  If the unit's run speed is more than 0 and the target is within the run action's range, it will run instead of walk.
Notifications - called from javascript.  You request orders with a target entity, order type, and whether the previous listeners you have requested for this entity should be destroyed.  When the target entity processes an order that was requested, that order is pushed onto the queue of the requester.  This is useful for things such as follow that require the actual order to perform the action.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3329 3db68df2-c116-0410-a063-a993310a9797