# As requested by Jason, now only one tab list in the Status Pane will be open at any time (hovering one will close any others that are open).
(Probably. I couldn't test with units due to the ModelRenderer javelinist crash, but it seems to have the desired effect for the Civ Centre. Which isn't the most conclusive test since the research list is empty anyway.)

# Implemented Andrew's formation functions in the formation "pick list".

(Again, probably. Since this is completely untestable without being able to select units, I can only speculate that it might work. The game didn't crash and burn on startup as a result of my meddling, which I take as a good sign.)

(Andrew: The changed section is a formation case switch at line 376 of functions_page_session_status_commands.js. I endeavoured to use all your functions to try them out. Like I said, I'm not sure if any of this works -- and being sick and feverish doesn't help -- but it might at least give you somewhere to start.)

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3723 3db68df2-c116-0410-a063-a993310a9797