#Bar borders and angle penalty
-Changes to notifications. They take different parameters now-see template_entity_script.js.  You can choose to destroy the notifiers yourself in the script (useful for idle)
-Added "idle" event with registerIdle and registerDamage to assist with the angle penalty.
-Bar border stuff
-Angle penalty is set up but untested-it just needs to use this.getAttackDirections() to find the number of directions the entity is being attacked from.  The penalty is specified in template_unit

There is a problem when the game exits-it attempts to destroy the notifiers in entity.cpp's constructor, where it calls DestroyAllNotifiers().  The problem is that the notifiers don't exist any longer because they've been destroyed. I would fix it but I'm leaving for vacation (Jason told me it was OK to comitt). Hope it isn't too much of a problem.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3732 3db68df2-c116-0410-a063-a993310a9797