Fixed a crash that happened for me in release mode when pressing the shift key. In Hotkey.cpp:594, there was a line like
	if( !unified[(*itKey)-UNIFIED_SHIFT] ) accept = false;

This was causing an index out of bounds when the key at *itKey was AND'ed with the HOTKEY_NEGATION_MASK and thus became greater than 65536 and way out of bounds of the array. For now I added a check that *itKey < HOTKEY_NEGATION_MASK, but maybe this is not the intended usage.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4111 3db68df2-c116-0410-a063-a993310a9797