#Bugfix: Shadows appeared to be rendered one frame too late close #141
The problem was that the shadow matrices were setup in BeginFrame, but GraphicsView calls SetCamera after BeginFrame. The solution was to move the shadow matrix setup into CRenderer::SetCamera.

This caused a second problem because RenderReflections/RenderRefractions used SetCamera to change the OpenGL matrices. Solved this problem by distinguishing explicitly between the camera used for the scene as a whole and the camera used to configure OpenGL. The latter can be a virtual camera for shadow or reflection rendering and similar render-to-texture effects.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4330 3db68df2-c116-0410-a063-a993310a9797