# Fixed walking on a multistep path, which sometimes stopped after the first few steps.
The problem was incorrect logic in the start-next-segment-of-the-path code that was added to allow multiple segments to be handled within the same update. I'm not sure how this worked before but it seemed to give continuous paths! Likely that's because the pathing for contact actions is different.

Also added more sanity checks in Aura.cpp for deleted entities in the m_Influenced queue. There was a crash caused by something here though it's hard to reproduce.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5597 3db68df2-c116-0410-a063-a993310a9797