add cache dir but ignore its contents.
(game code mounts data/cache into VFS. the mount call assumes the directory exists (for security reasons we do not want to create arbitrary directories). if not (as recently reported by matei), an error in wfilesystem:247 will result. adding cache/ to SVN ensures it will be present in fresh checkouts; existing checkouts should not be affected.)

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5996 3db68df2-c116-0410-a063-a993310a9797