Fix rendering artifacts with novbo=true gentangents=true
glDrawRangeElements needs to know the range of vertexs that are used
by the index array. With VBOs it doesn't really matter if the range
is wrong (all the vertexes are in GPU memory anyway), but with CPU
vertex arrays the driver has to memcpy the given range of data, so
incorrect bounds will result in garbage data being rendered.

With gentangents, the rendered mesh can have more vertexes than the
original CModelDef, but was rendered with the CModelDef's vertex count.

Use the correct vertex count instead.

Refs #2050.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@13734 3db68df2-c116-0410-a063-a993310a9797