Unify script conversions and remove JSInterface_Vector3D.
Because it was historically grown, we have some duplicated code for converting script types to native types. 
This patch removes the file JSConversions.cpp and moves some code to ScriptConversions.cpp. 
The places using JSConversions.cpp are changed to use the ScriptInterface's conversion functions in ScriptConversions.cpp.
I also removed JSInterface_Vector3D because it had additional requirements to the conversion code that no other code has and because it's currently not used. I think it doesn't make sense to maintain code just because it could possibly be used again later.  

Closes #2213
Refs #1886


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@14036 3db68df2-c116-0410-a063-a993310a9797