Workaround to reduce the impact of AI performance problems when planning new bases.
When no free territory was left on the map, the AI tried to place a CC every 10th AI turn.
That's about all 10 seconds and it freezed the game for about a second.
Now when you played against 4 players it freezed aobut 4 seconds every 10 seconds, which is very bad.
I wanted to solve this properly but realized that I'd have to change a lot in the terrain analysis code that will probably be replaced anyway when we have a new pathfinder with a proper interface.
Also removes an unused variable (creating a territory map and not using it can't be good for performance either).



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@14859 3db68df2-c116-0410-a063-a993310a9797