Works around infinite recursion problem.
This problems occurs with the new SpiderMonkey in debug mode, but it's not a SpiderMonkey issue.
Unfortunately I couldn't find a clean solution, so I just documented the problem in an extensive comment block and applied a workaround.
A clean solution would probably require changing the approach for counting allocations completely.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@14874 3db68df2-c116-0410-a063-a993310a9797