Internationalization of the C++ side
Provides the logic to detect the system language, load translations, and use
loaded translations both in the C++ and the JavaScript side.

This patch includes code by Yves, sanderd17, leper, historic_bruno and
Gallaecio. It’s worth noting that Yves and historic_bruno were also the main
contributors behind the changes in r14951 as well.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@14953 3db68df2-c116-0410-a063-a993310a9797