Attempt to work around LOS texture alignment bug.
For unknown reasons, sometimes glTexSubImage2D acts as if GL_UNPACK_ALIGNMENT = 2 or 4, instead of 1. This causes the odd-sized array uploaded for the LOS texture to be interpreted incorrectly, and the LOS texture gets rendered very incorrectly.

Pad the array to a multiple of 4 in all cases, so that GL_UNPACK_ALIGNMENT shouldn't affect it.

Hopefully fixes #2594.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15216 3db68df2-c116-0410-a063-a993310a9797