New unit renderer.
Instead of each CCmpVisualActor rendering itself individually, collect all the units in a single CCmpUnitRenderer. This avoids the overhead of doing Interpolate/RenderSubmit calls every frame for every object in the world. It also allows more efficient culling.

CCmpUnitRenderer knows the positions of each object at the start and end of each turn, and computes the bounding sphere of the object along that path. That allows quick culling without recomputing the precise interpolated transform every frame. (In the future it could be improved much more.)

Clarify and clean up the sending of PositionChanged messages, and add new InterpolatedPositionChanged.

Remove the forceFloating parameter from GetInterpolatedTransform, since it doesn't fit the new design. Replace it with a (non-synchronised) flag in CCmpPosition.

Move construction progress from CCmpVisualActor to CCmpPosition, so that it consistently affects all position/transform computation.

Refs #2337.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15265 3db68df2-c116-0410-a063-a993310a9797