Fix units falling off the edge of the world.
I guess changes to the map loading sequence caused the TerrainChanged message to be sent before the map was switched from square to circular instead of after. The pathfinder didn't notice the switch, so it continued treating the map as if it were square, allowing units to walk into the permanent map-corner SOD and vanish, and allowing territories to expand into the SOD.

Tell the pathfinder explicitly when the map shape changes, so it can discard its cached data correctly.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15277 3db68df2-c116-0410-a063-a993310a9797