Fix culling for shadows and reflections.
Previously we had a single culling frustum based on the main camera, and any object outside the frustum would never get rendered, even if it should actually contribute to shadows or reflections/refractions. This caused ugly pop-in effects in the shadows and reflections while scrolling.

Extend the renderer to support multiple cull groups, each with a separate frustum and with separate lists of submitted objects, so that shadows and reflections will render the correctly culled sets of objects.

Update the shadow map generation to compute the (hopefully) correct bounds and matrices for this new scheme.

Include terrain patches in the shadow bounds, so hills can cast shadows correctly.

Remove the code that tried to render objects slightly outside the camera frustum in order to reduce the pop-in effect, since that was a workaround for the lack of a proper fix.

Remove the model/patch filtering code, which was used to cull objects that were in the normal camera frustum but should be excluded from reflections/refractions, since that's redundant now too.

Inline DistanceToPlane to save a few hundred usecs per frame inside CCmpUnitRenderer::RenderSubmit.

Fixes #504, #579.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15445 3db68df2-c116-0410-a063-a993310a9797