Makes tinygettext handle missing plural forms more sanely.
Print an error to the console and return the original (untranslated) string instead of triggering an assertion in debug builds or a crash and memory corruption in release builds.
Works around a crash on the history screen. Why the plural form is missing is another topic and will have to be solved separately.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15493 3db68df2-c116-0410-a063-a993310a9797