Hide changes to buildings in the fog-of-war.
To achieve this, mirage entities are created per player, to replace the real entities when these ones fall into the fog-of-war. These mirage entities are created on-the-fly, and destroyed when they get back in sight.
This depends heavily on the VisibilityChanged message added in r15508.

As a temporary adjustment, territories do not explore the map anymore when their borders change. See #2709.

Fixes #599

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15612 3db68df2-c116-0410-a063-a993310a9797