Prevent copying of ScriptInterface objects.
Copying a ScriptInterface is not allowed. The destructor destroys the JSContext and if there's a copy around that still uses it, that causes crashes which can be quite hard to debug. This change prevents copying by of ScriptInterface objects by making the copy constructor private, which will reveal such errors at compile time.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15625 3db68df2-c116-0410-a063-a993310a9797