Reverts r15669.
It doesn't compile on Windows and has some fundamental problems that need to be sorted out first. The Windows GUI can take several arguments (-minizimed, -keep and -title) and it needs to extract them from argv somehow. There's a similar problem with the QT GUI and even if we don't need it, it has to be fixed if we want to submit the patch upstream. IMO it's better to revert this and live with the problems of #2488 at the moment than implementing some hacks to make it compile and introcing additional differences to cxxtest that can't be committed upstream yet.

Refs #2488


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15683 3db68df2-c116-0410-a063-a993310a9797