Tunes GC scheduling a bit to reduce memory usage.
The main problem was that GC was only called from the simulation before this patch. This means when you were waiting in the multiplayer lobby or just had the GUI open, it only called GC when getting close to the JS runtime size limit (I assume). Another problem was the Net Server runtime which didn't GC either. Here the runtime size limit is 16 MB though, so it's not too terrible. These issues have both been addressed and GC has been given a bit more time per incremental slice to make sure it gets done in time. It's still far from perfect, but there are too many changes in SpiderMonkey related to GC, so I don't want to spend too much time on this yet. 

Refs #2808   


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15787 3db68df2-c116-0410-a063-a993310a9797