Modify GC scheduling and reduce VFS cache size.
It seems like there is a memory leak if we haven't finished with the marking phase of an incremental GC and SpiderMonkey has to trigger a full GC because it runs out of memory. With this patch we stop trying to make incremental GCs if we are above 1/2 of the runtime size and do Full GCs instead. This should make such low memory conditions even less likely than they were already after the previous patch. Also reduce the maximum VFS cache size to 400 MB.
Refs #2808


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15831 3db68df2-c116-0410-a063-a993310a9797