Split ScriptRuntime and ScriptInterface code to separate files.
The runtime is becoming more and more important in the JSAPI. As a result, we also have more functionality on the runtime level and having the whole ScriptRuntime class hidden in ScriptInterface.cpp doesn't make sense anymore. ScriptInterface.cpp also has become quite a large file and pulling out the runtime part makes it a bit smaller.

Refs #2462 


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@15961 3db68df2-c116-0410-a063-a993310a9797