Add a new spatial subdivision, based on an old patch by wraitii.
This subdivision is faster but less precise, so range queries get more entities and are a bit slower (up to 1ms approx.), but the overall gain on a simulation update is always positive and can reach 10ms per frame.

For now, this new subdivision is only used by the range manager, integrating it in the obstruction manager might be sensible.

Refs #2430

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16540 3db68df2-c116-0410-a063-a993310a9797