Change the handling of modified entities in the visibility update.
The game has to deal with situations such as: the visibility of an entity changes, a mirage is created for it -> the mirage visibility is updated -> the entity visibility is updated back.
All of this process now happens in the same turn, and all updates are guaranteed to be performed. This fixes a source of serialization errors and rejoin OOSes.

Fixes #3107

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16857 3db68df2-c116-0410-a063-a993310a9797