Fix the OOS reported by elexis in #3335, and clean some whitespace and style.
The list of modified entities was thrown out on global visibility updates (those happen on rejoin) but only in-world entities were actually updated then, causing problems with garrisoning.

Now the list of modified entities can happen to be large, so replace the hacky check for infinite loops by a real check.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16962 3db68df2-c116-0410-a063-a993310a9797