Based on previous experimental changes, major update to the unit motion.
With this change, units will not check their movement against all obstructions when moving: terrain and static obstructions are assumed to be handled by the long-range pathfinder.
However, when static obstructions are changed, the paths have to be invalidated. In order to minimize the performance impact, units will check for obstructions when they move after a passability change. If they collide with something, they will recompute a path that will take into account the new passability map.

Also includes some code cleanup. This patch should not change performance a lot: the lower number of checks should give a small performance improvement while using the message broadcasting system should hurt it a bit.

Fixes #3376, #3337, #1914.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16998 3db68df2-c116-0410-a063-a993310a9797