Use let instead of var in the gamesetup.
Doing so follows the principle of least exposure and means those variables will be destroyed when leaving the block (i.e. memory being freed up earlier).

The blocks affected by this commit are short (mostly 3-5 lines).
The variables are not referenced before or after the block, nor in their temporal dead zones.




git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@17424 3db68df2-c116-0410-a063-a993310a9797