Gamesetup cleanup (simplify GUI calls)
Make a hidden function globally available and use it everywhere to simplify the hiding/showing of GUI elements.
Use strings instead of the references to the actual GUI objects, so that the code doesn't rely on variables anymore.
Use a default argument so we can reuse it with regards to other conditions like scenario maps or multiplayer.
Notice the JS variables have the same name as the GUIObjects, so the replacement is simple.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@17447 3db68df2-c116-0410-a063-a993310a9797