Replace the inscription on the altar in the main menu background.
The new inscription fits better with the scene, set in Sparta, and is the Greek original epitaph written on the Cenotaph of Thermopylae, according to Herodotus:
"Go, tell the Spartans, stranger passing by, that here, obedient to their laws, we lie."

The previous inscription was apparently a fraction of a sentence and was about Spartan foreign policy in Ionia. 

Suggestion and research by Danny, art by wowgetoffyourcellphone.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@17805 3db68df2-c116-0410-a063-a993310a9797