Win and defeat cleanup, fixes #4013.
Add a new simulation message and chat notification for players who won.
Avoid duplicate playerstate messages in the sim and GUI by triggering changes with a function instead of a message.
Reveal the map on defeat/win exclusively in the player component (instead of having a silly GUI proxy and doing it also in the EndGameManager sometimes).
Remove the skipAlliedVictory argument from the player component, since that shouldn't contain references to the EndGameManager, refs #3970.

Show a proper message box on win/loss and add the hint for hosts disconnecting other players.
Do defeat/win procedure in the GUI when such a message arrives, instead of checking onTick for playerstate changes.
Thus don't show that confirmation again on every rejoin.
Don't show a pointless message box if IsAtlasRunning.

Explain that the session.js variable is needed to avoid an order-of-execution bug.
Select "observer" item when rejoining as a defeated player.
Remove an unneeded call to updateTopPanel.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@18441 3db68df2-c116-0410-a063-a993310a9797