Ram garrison balancing.
Garrison up to 10 men in rams and 12 in persian ones, in order to
have a greater chance of attacking enemies with many melee champs defending,
since the new balancing in alpha 21 requires siege engines to destroy buildings, see r18037 and r18045.

Remove undocumented garrisoned healrate of rams introduced in r11292.
That commit introduced it in many places, which were removed in later commits.

Leave the positive garrisoned healrate of ships and special buildings for now,
assuming that there are supplies on ships (but not in rams).

Add garrisoned healrate tooltips in the following commit.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@18528 3db68df2-c116-0410-a063-a993310a9797