Make wooden tower upgradable to the defense tower. Reviewed by fatherbushido, siole and Hannibal Barca, fixes #4125.
Let towers cost 50 wood and reduce the stone cost by 50.
The upgrade costs so much that it's cheaper to build a stone tower than building and upgrading a wooden tower, but still cheaper than building a wooden and stone tower.
Increase the minimum distance between wooden tower to 60, so that the upgrade won't circumstance the minDistance of the stone tower.
Add missing newline to the entity limit tooltip of the upgrade component (refs #2706) that was never visible before.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@18754 3db68df2-c116-0410-a063-a993310a9797