Remove a hunk with commented out code and a hunk of broken duplicate code from survival of the fittest, refs #3102.
The removed code is redundant with the AlliedVictoryCheck of the EndGameManager which is called for all victory conditions.
The code is broken because it suggests last man standing mode while actually not implementing it.
As of #4143, players can chose between allied victory or last man standing mode. They should be able to chose on this map too.



git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@18906 3db68df2-c116-0410-a063-a993310a9797