Fix inconsistent disabled resign and pause button
Summary:
The pause button displays an additional GUI element on top of the button containing the text.
This apparently unneeded element uses a style forcing the button text to be white, even if the element is disabled.

Test Plan:
Start a multiplayergame with all players in the same team, winning it instantly.
The resign and pause button become disabled, but have different text colors.
Apply the patch and notice the issue is gone and that the XML is more consistent.
Try pausing to see that the pause button caption still updates.
Convince yourself that there are no other occurances of `pauseButtonText`.


Reviewers: fatherbushido

Reviewed By: fatherbushido

Subscribers: Vulcan

Differential Revision: https://code.wildfiregames.com/D91

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@19173 3db68df2-c116-0410-a063-a993310a9797