Implement shared vision at the entity level
Summary:
To fix tickets like #3335, having a shared vision at the entity level is needed. This patch implements that in CCmpRangeManager, interfaced with a new JS component VisionSharing to manage the shared entities visions. As an example of use case (in addition to garrisoning in allied structure without the sharedLos tech), we can think of bribing enemy units: there is a part about it in the patch, but this one is very wip and not intended to be committed but rather for test purposes of the feature.
So when garrisoning a unit in another player's building, the vision of the garrisonHolder is shared (part intended for review). In addition, for tests of the feature, when clicking on the new bribe icon in the diplomacy window, a random unit of the chosen player is bribed and share its vision during 15 s.

Test Plan: Garrison a unit in an allied structure without the sharedLos tech, or test the wip bribe feature from the diplomacy window.

Reviewers: Itms

Reviewed By: Itms

Subscribers: Stan, leper, O11 Templates, wraitii, elexis, fatherbushido, Itms, Vulcan, O1 C++ Simulation

Differential Revision: https://code.wildfiregames.com/D60

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@19175 3db68df2-c116-0410-a063-a993310a9797